+++ .env.example
@@ -11,4 +11,4 @@ DEBUG=TRUE
 
 # Front-End Variables
 BASENAME=/
-#BACKEND_URL=
+BACKEND_URL=https://super-duper-invention-wr77xr77vxq52vvwp-3001.app.github.dev/
