+++ package.json
@@ -80,10 +80,11 @@
 		"@fortawesome/free-solid-svg-icons": "^6.4.2",
 		"@fortawesome/react-fontawesome": "^0.2.0",
 		"babel-plugin-macros": "^3.1.0",
+		"bootstrap": "^5.3.2",
 		"cloudinary-core": "^2.13.0",
 		"cloudinary-react": "^1.8.1",
-		"bootstrap": "^5.3.2",
 		"cors": "^2.8.5",
+		"fontawesome": "^5.6.3",
 		"prop-types": "^15.6.1",
 		"react": "^16.8.4",
 		"react-dom": "^16.8.4",
