+++ src/api/models.py
@@ -8,6 +8,12 @@ class User(db.Model):
     name = db.Column(db.String(120), unique=False, nullable=False)
     email = db.Column(db.String(120), unique=True, nullable=False)
     password = db.Column(db.String(80), unique=False, nullable=False)
+    firstName = db.Column(db.String(80), unique=False, nullable=True)
+    lastName = db.Column(db.String(80), unique=False, nullable=True)
+    userName = db.Column(db.String(80), unique=False, nullable=True)
+    linkedIn = db.Column(db.String(200), unique=False, nullable=True)
+    github = db.Column(db.String(200), unique=False, nullable=True)
+    avatar = db.Column(db.String(200), unique=False, nullable=True)
 
     def __repr__(self):
         return f'<User {self.email}>'
@@ -17,6 +23,10 @@ class User(db.Model):
             "id": self.id,
             "name": self.name,
             "email": self.email,
+            "avatar": self.avatar,
+            "firstName": self.firstName,
+            "lastName": self.lastName,
+            "linkedIn": self.linkedIn,
+            "github": self.github,
             # do not serialize the password, its a security breach
         }
-
